pyproject.toml
[project]
name = "retrieval"
version = "0.1.0"
description = "Lab 3 scaffold to P2 document retrieval system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.21",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.13",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "--cov=src/retrieval",       # always look for coverage just in these files
    "--cov-report=term-missing", # always report which code lines are uncovered (instead of only count)
    "--cov-report=html",         # make the report pretty to look at
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]    # Just errors, pyflakes, and imports
ignore = ["E501"]           # Formatter handles line length

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F401"]        # Allow unused imports in tests
